FastApiBackend/requirements.txt
fastapi==0.103.2
numpy~=1.26.0
Pillow==10.0.1
tensorflow==2.14.0

starlette~=0.27.0